Fix non-intuitive and kind of stupid issue that lead to https://github.com/fake-name/PyGalil/issues/7.
Basically, `inMotion()` returns a literal string of either `1.0000` for true, or `0.0000` for false. Since `bool('0.0000')` == True, this means /using/ the output of `inMotion()` will inevitably require a separate manual cast. Since the response of this query is only ever boolean, move the cast to a bool into the actual call so it returns True/False like a sensible API should.